Refactor entry fixtures
