+++ p2panda-rs/Cargo.toml
@@ -23,6 +23,7 @@ dwarf-debug-info = false
 [[bin]]
 name = "cbor-test-data"
 path = "src/test_utils/generate_test_data.rs"
+required-features = ["testing"]
 
 [lib]
 crate-type = ["cdylib", "rlib"]
