+++ p2panda-rs/src/lib.rs
@@ -75,6 +75,7 @@ pub mod operation;
 pub mod schema;
 pub mod secret_group;
 pub mod storage_provider;
+#[cfg(any(feature = "testing", test))]
 pub mod test_utils;
 #[cfg(target_arch = "wasm32")]
 pub mod wasm;
