+++ p2panda-rs/src/test_utils/fixtures/fixtures.rs
@@ -8,7 +8,7 @@ use crate::identity::KeyPair;
 use crate::operation::{OperationEncoded, OperationValue};
 use crate::schema::SchemaId;
 use crate::test_utils::constants::TEST_SCHEMA_ID;
-use crate::test_utils::fixtures::{create_operation, entry, seq_num};
+use crate::test_utils::fixtures::{create_operation, entry, log_id, seq_num};
 use crate::test_utils::utils;
 
 /// Fixture struct which contains versioned p2panda data for testing.
@@ -48,7 +48,7 @@ pub fn v0_3_0_fixture() -> Fixture {
         entry_signed_encoded: EntrySigned::new("009cdb3a8c0c4b308173d4c3c43a67a6d013444af99acb8be6c52423746d9aa2c10101bc00207597aa680a7f619d72ec4410bb3a0af4bcb66509e43c1ddec70beefd4b158f5c1a0836975a8a23d92e7ff3d23742dfb4a5c447b2ef7b86fd1063743ba6bb50e0ddceff7d3814825aaf35cf1d2288061fcfff00375b91dcfc38f945a798d1810a").unwrap(),
         operation_encoded: OperationEncoded::new("a466616374696f6e6663726561746566736368656d61784a76656e75655f30303230633635353637616533376566656132393365333461396337643133663866326266323364626463336235633762396162343632393331313163343866633738626776657273696f6e01666669656c6473a26b6465736372697074696f6ea26474797065637374726576616c756571666f7220706c6179696e67206368657373646e616d65a26474797065637374726576616c7565656368657373").unwrap(),
         key_pair,
-        entry: entry(operation, seq_num(1), None, None)
+        entry: entry(operation, seq_num(1), None, None, log_id(1))
     }
 }
