+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/mod.rs
@@ -1,0 +1,3 @@
+mod entry_fixtures;
+
+pub use entry_fixtures::*;
