+++ p2panda-rs/src/test_utils/mod.rs
@@ -5,6 +5,7 @@
 //! It includes fixtures and templates which can be injected into tests, mock node and client
 //! implementations, methods for generating test data (used in `p2panda-js`).
 pub mod constants;
+#[cfg(test)]
 pub mod fixtures;
 pub mod mocks;
 pub mod test_data;
